+++ bonus/src/objects/cone_bonus.c
@@ -6,7 +6,7 @@
 /*   By: ymiao <ymiao@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/31 20:19:26 by jinhuang          #+#    #+#             */
-/*   Updated: 2025/09/02 20:55:44 by ymiao            ###   ########.fr       */
+/*   Updated: 2025/09/05 01:06:53 by ymiao            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 static void	fill_co_info(t_cone *co, t_ray *ray, t_co_info *info)
 {
-	double	cos2;
+	float	cos2;
 
-	cos2 = cos(co->angle) * cos(co->angle);
+	cos2 = cosf(co->angle) * cosf(co->angle);
 	info->a = info->dv * info->dv - cos2;
 	info->b = 2 * (info->dv * info->ocv
 			- cos2 * vector_dot(ray->direction, info->oc));
@@ -25,12 +25,12 @@ static void	fill_co_info(t_cone *co, t_ray *ray, t_co_info *info)
 	info->discr = info->b * info->b - 4 * info->a * info->c;
 }
 
-static double	hit_cone_body(t_cone *co, t_ray *ray)
+static float hit_cone_body(t_cone *co, t_ray *ray)
 {
-	double		t1;
-	double		t2;
-	double		m1;
-	double		m2;
+	float		t1;
+	float		t2;
+	float		m1;
+	float		m2;
 	t_co_info	info;
 
 	info.oc = vector_sub(ray->origin, co->center);
@@ -38,29 +38,29 @@ static double	hit_cone_body(t_cone *co, t_ray *ray)
 	info.ocv = vector_dot(info.oc, co->normal);
 	fill_co_info(co, ray, &info);
 	if (info.discr < 0)
-		return (-1.0);
-	t1 = (-info.b - sqrt(info.discr)) / (2 * info.a);
+		return (-1.0f);
+	t1 = (-info.b - sqrtf(info.discr)) / (2 * info.a);
 	m1 = info.ocv + t1 * info.dv;
 	if (t1 <= EPSILON || m1 < 0 || m1 > co->height)
-		t1 = -1.0;
-	t2 = (-info.b + sqrt(info.discr)) / (2 * info.a);
+		t1 = -1.0f;
+	t2 = (-info.b + sqrtf(info.discr)) / (2 * info.a);
 	m2 = info.ocv + t2 * info.dv;
 	if (t2 <= EPSILON || m2 < 0 || m2 > co->height)
-		t2 = -1.0;
+		t2 = -1.0f;
 	if (t1 > 0 && t2 > 0)
-		return (fmin(t1, t2));
+		return (fminf(t1, t2));
 	if (t1 > 0)
 		return (t1);
 	return (t2);
 }
 
-static double	hit_cone_cap(t_cone *co, t_ray *ray)
+static float hit_cone_cap(t_cone *co, t_ray *ray)
 {
 	t_plane		cap;
 	t_vector	base;
 	t_vector	p;
-	double		t;
-	double		r;
+	float		t;
+	float		r;
 
 	base = vector_add(co->center,
 			vector_mult(vector_normalize(co->normal), co->height));
@@ -70,17 +70,17 @@ static double	hit_cone_cap(t_cone *co, t_ray *ray)
 	if (t > 0)
 	{
 		p = vector_add(ray->origin, vector_mult(ray->direction, t));
-		r = tan(co->angle) * co->height;
+		r = tanf(co->angle) * co->height;
 		if (vector_length_sq(vector_sub(p, base)) <= r * r)
 			return (t);
 	}
-	return (-1.0);
+	return (-1.0f);
 }
 
-double	hit_cone(t_cone *co, t_ray *ray)
+float hit_cone(t_cone *co, t_ray *ray)
 {
-	double	t_body;
-	double	t_cap;
+	float	t_body;
+	float	t_cap;
 
 	t_body = hit_cone_body(co, ray);
 	t_cap = hit_cone_cap(co, ray);
@@ -93,16 +93,16 @@ t_vector	cone_normal_at(t_cone *co, t_vector point)
 {
 	t_vector	v;
 	t_vector	oc;
-	double		m;
-	double		k;
+	float		m;
+	float		k;
 	t_vector	normal;
 
 	v = vector_normalize(co->normal);
 	oc = vector_sub(point, co->center);
 	m = vector_dot(oc, v);
-	if (fabs(m - co->height) < EPSILON)
+	if (fabsf(m - co->height) < EPSILON)
 		return (v);
-	k = tan(co->angle);
+	k = tanf(co->angle);
 	normal = vector_sub(oc, vector_mult(v, m * (1 + k * k)));
 	if (vector_length_sq(normal) < EPSILON)
 		return (vector_mult(v, -1));
