+++ bonus/src/render/camera_bonus.c
@@ -6,7 +6,7 @@
 /*   By: ymiao <ymiao@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/16 23:05:09 by ymiao             #+#    #+#             */
-/*   Updated: 2025/09/02 19:22:44 by ymiao            ###   ########.fr       */
+/*   Updated: 2025/09/05 01:13:10 by ymiao            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,18 +22,18 @@
 void	setup_cam_coords(t_camera *camera)
 {
 	t_vector	world_up;
-	double		theta;
+	float		theta;
 
-	camera->aspect_ratio = (double)WIDTH / (double)HEIGHT;
+	camera->aspect_ratio = (float)WIDTH / (float)HEIGHT;
 	world_up = vector_init(0, 1, 0);
 	if (fabs(vector_dot(camera->direction, world_up)) > 0.999)
 		world_up = vector_init(0, 0, 1);
-	camera->w = vector_mult(camera->direction, -1.0);
+	camera->w = vector_mult(camera->direction, -1.0f);
 	camera->w = vector_normalize(camera->w);
 	camera->u = vector_cross(world_up, camera->w);
 	camera->u = vector_normalize(camera->u);
 	camera->v = vector_cross(camera->w, camera->u);
-	theta = camera->fov * M_PI / 180.0;
-	camera->half_height = tan(theta / 2.0);
+	theta = camera->fov * M_PI / 180.0f;
+	camera->half_height = tanf(theta / 2.0f);
 	camera->half_width = camera->half_height * camera->aspect_ratio;
 }
