+++ src/parsing/parsing.h
@@ -6,7 +6,7 @@
 /*   By: ymiao <ymiao@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/31 03:17:29 by ymiao             #+#    #+#             */
-/*   Updated: 2025/09/02 18:59:56 by ymiao            ###   ########.fr       */
+/*   Updated: 2025/09/05 01:06:52 by ymiao            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ t_color		parse_color(char *str);
 t_vector	parse_vector(char *str);
 
 void		free_tokens(char **tokens);
-double		ft_atof(const char *str);
+float		ft_atof(const char *str);
 int			count_tokens(char **tokens);
 bool		is_normalized_vector(t_vector v);
 void		trim_newline(char *str);
